created new components for better structuring styled new components created pages about contact to be used for about and contact
