+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^12.8.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "^5.0.1",
     "react-toastify": "^8.1.0",
     "web-vitals": "^1.1.2"
