+++ src/App.css
@@ -5,50 +5,3 @@
   justify-content: center;
   background-color: #000000;
 }
-
-.generator {
-  background-image: linear-gradient(-225deg, #430c0f 0%, #2d2387 100%);
-  border-radius: 10px;
-  box-shadow: 0px 2px 10px rgba(245, 244, 244, 0.863);
-  padding: 30px;
-}
-.generator__header {
-  text-align: center;
-  color: rgb(244, 244, 245);
-  margin-bottom: 20px;
-  font-size: 30px;
-}
-.generator__password {
-  background-color: #d50b12;
-  padding: 13px 10px;
-  color: #fff;
-  margin-bottom: 15px;
-  height: 46px;
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 50px;
-}
-.copy__btn {
-  color: #fff;
-  background: #d50b12;
-  border: none;
-  cursor: pointer;
-}
-.generator__btn {
-  background: #d50b12;
-  border: none;
-  width: 100%;
-  margin-top: 10px;
-  padding: 25px;
-  color: #fff;
-  border-radius: 30px;
-  font-size: 17px;
-  cursor: pointer;
-  box-shadow: -2px 2px 10px rgba(245, 244, 244, 0.863);
-}
-.form-group {
-  display: flex;
-  justify-content: space-between;
-  color: whitesmoke;
-  margin-bottom: 25px;
-}
