+++ src/App.js
@@ -1,175 +1,16 @@
-import React, { useState } from "react";
+import React from "react";
 import "./App.css";
-import {
-  numbers,
-  upperCaseLetters,
-  lowerCaseLetters,
-  specialCharacters,
-} from "./Character";
-import { toast, ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
-import { COPY_Fail, COPY_SUCCESS } from "./message";
 import Header from "./Header";
 import Footer from "./Footer";
+import Passwordgenerator from "./Passwordgenerator";
 
 const App = () => {
-  const [password, setPassword] = useState("");
-  const [passwordLength, setPasswordLength] = useState(26);
-  const [includeUpperCase, setIncludeUpperCase] = useState(false);
-  const [includeLowerCase, setIncludeLowerCase] = useState(false);
-  const [includeNumbers, setIncludeNumbers] = useState(false);
-  const [includeSymbols, setIncludeSymbols] = useState(false);
-  const handleGeneratePassword = () => {
-    if (
-      !includeUpperCase &&
-      !includeLowerCase &&
-      !includeNumbers &&
-      !includeSymbols
-    ) {
-      notify("To generate password you must select atleast one checkbox", true);
-    } else {
-      let characterList = "";
-      if (includeNumbers) {
-        characterList = characterList + numbers;
-      }
-      if (includeUpperCase) {
-        characterList = characterList + upperCaseLetters;
-      }
-      if (includeLowerCase) {
-        characterList = characterList + lowerCaseLetters;
-      }
-      if (includeSymbols) {
-        characterList = characterList + specialCharacters;
-      }
-      setPassword(createPassword(characterList));
-      notify("Password is generated successfully", false);
-    }
-  };
-  const createPassword = (characterList) => {
-    let password = "";
-    const characterListLength = characterList.length;
-    for (let i = 0; i < passwordLength; i++) {
-      const characterIndex = Math.round(Math.random() * characterListLength);
-      password = password + characterList.charAt(characterIndex);
-    }
-    return password;
-  };
-  const copyToClipboard = (password) => {
-    navigator.clipboard.writeText(password);
-  };
-  const notify = (message, hasError = false) => {
-    if (hasError) {
-      toast.error(message, {
-        position: "top-center",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-      });
-    } else {
-      toast(message, {
-        position: "top-center",
-        autoClose: 5000,
-        hideProgressBar: false,
-        closeOnClick: true,
-        pauseOnHover: true,
-        draggable: true,
-        progress: undefined,
-      });
-    }
-  };
-  const handleCopyPassword = (e) => {
-    if (password === "") {
-      notify(COPY_Fail, true);
-    } else {
-      copyToClipboard(password);
-      notify(COPY_SUCCESS);
-    }
-  };
-
   return (
-    <div className="App">
+    <div>
       <Header />
-      <div className="container">
-        <div className="generator">
-          <h2 className="generator__header">Password Generator</h2>
-          <div className="generator__password">
-            <h2>{password}</h2>
-            <button className="copy__btn">
-              <i onClick={handleCopyPassword} className="far fa-clipboard"></i>
-            </button>
-          </div>
-          <div className="form-group">
-            <label htmlFor="password-strength">Password Length</label>
-            <input
-              className="pw"
-              defaultValue={passwordLength}
-              onChange={(e) => setPasswordLength(e.target.value)}
-              type="number"
-              id="password-stregth"
-              name="password-strength"
-              max="30"
-              min="5"
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="uppercase-letters">Include Uppercase</label>
-            <input
-              checked={includeUpperCase}
-              onChange={(e) => setIncludeUpperCase(e.target.checked)}
-              type="checkbox"
-              id="uppercase-letters"
-              name="uppercase-letters"
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="lowercase-letters">Include Lowercase</label>
-            <input
-              checked={includeLowerCase}
-              onChange={(e) => setIncludeLowerCase(e.target.checked)}
-              type="checkbox"
-              id="lowercase-letters"
-              name="lowercase-letters"
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="include-numbers">Include Numbers</label>
-            <input
-              checked={includeNumbers}
-              onChange={(e) => setIncludeNumbers(e.target.checked)}
-              type="checkbox"
-              id="include-numbers"
-              name="include-numbers"
-            />
-          </div>
-          <div className="form-group">
-            <label htmlFor="include-symbols">Include Symbols</label>
-            <input
-              checked={includeSymbols}
-              onChange={(e) => setIncludeSymbols(e.target.checked)}
-              type="checkbox"
-              id="include-symbols"
-              name="include-symbols"
-            />
-          </div>
-          <button onClick={handleGeneratePassword} className="generator__btn">
-            Generate Password
-          </button>
-          <ToastContainer
-            position="top-center"
-            autoClose={5000}
-            hideProgressBar={false}
-            newestOnTop={false}
-            closeOnClick
-            rtl={false}
-            pauseOnFocusLoss
-            draggable
-            pauseOnHover
-          />
-        </div>
-      </div>
+
+      <Passwordgenerator />
+
       <Footer />
     </div>
   );
